switch window manager theme to Clearlooks-Phenix-Modified
I was using the Chicago95 theme for the window manager,
because I was nostalgic about the Windows 98 and wanted
to have the same experience on my GNU/Linux system. But
the problem was that the Chicago95 for GTK2 was not good
and barely usable, so I stuck with using
Clearlooks-Phenix-Modified theme for GTK2 and Chicago95
for the window manager theme (which draws the window frame).
And with time it became clear that they were not meant for
each other, they were disconnected, out of balance and the
design was not uniform or unified.

So, I tried again the Clearlooks-Phenix-Modified theme
for the window manager and it felt much better. The system
seems to be unified and in sync again.

Now, I think I'm not using anything more from Chicago95 theme.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>