bash: put present working directory into a separate variable
The \w variable displays the current (or present) working directory
with full path. I think this will help with some color changes I want
to make.

PWD is the acronym for present working directory and was chosen because
there is a command line tool in coreutils with the same name.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>