bash: remove the empty line between host and pwd
The empty line was used to add an empty space between
the hostname name and the pwd. With the previous changes
this change became more possible. Also the ${RESET} was
moved to the end of the BP_HOST, which is a pattern we
may apply to the other BP_XXX variables.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>