bash: show hostname in the prompt string again
Previously, the hostname was hidden by commit 8f77c49788d8 ("hide hostname
when not in an ssh connection"). Back then, the hostname was "apollo", but
now it is "lp2" which is much shorter, it is only 3 characters long. So,
together with "@" it takes only 4 characters on the screen. Another reason is
that the rules to hide the username are complex and I couldn't get around
it yet.

There is one more reason for restoring the hostname and always showing it in
the prompt string. The reason is that it brings us closer to the standard
string used in Ubuntu, which is "\u@\h:\w\$ ". The default one in Fedora
is "[\u@\h \W]\\$ ".

So, instead of wasting too much time trying to figure out how to hide the
username as well (at least for now), let's put the hostname back.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>