bash: show only the last component of the working directory
Instead of showing the full path of the working directory, the \W will
show only the basename of the working directory.

Take a look at the before and after screenshots to see the result.

This is what the default prompt in Fedora 11 does.

I think the main advantage of doing this way is that when going deep in
a directory hierarchy, the space for writing commands will not decrease
dramatically. It will always be short, unless you have a big directory name.

If you forget were you are and want to know the full path, you just type "pwd".

Signed-off-by: Thiago Farina <tfransosi@gmail.com>