bash: change current working directory color back to blue
This is for a fancy prompt with _colors_, so having the working directory
in blue color is not that bad.

Blue is also the default color used in Ubuntu's default prompt. So, with this
we get back closer to its default prompt string.

For future reference, this is the prompt string in Ubuntu from .bashrc:

PS1='${debian_chroot:+($debian_chroot)}\[\033[01;32m\]\u@\h\[\033[00m\]:\[\\
033[01;34m\]\w\[\033[00m\]\$ '

Signed-off-by: Thiago Farina <tfransosi@gmail.com>