bash: do not use '>' as a prompt symbol
In commit 1fd6d35eeb3d ("change the prompt symbol to ">"")
I changed the prompt symbol from '$' to '>' because I wanted
to make my prompt look similar to the Command Prompt of Windows.
But I didn't know at the time reasons not to use '>'.

Now I know a few reasons why '>' should not be used as a
prompt symbol in the shell running under terminal emulators on
UNIX systems like GNU/Linux.

Probably, the main one, is that in Bash shell '>' is a redirection
operator, that refers to the standard output (file descriptor 1) [0].

I just learned from [1] that using '>' sign as a prompt symbol can
have disastrous consequences.

If the prompt with '>' is copy/pasted by accident it can cause a
redirection and overwite the contents of file, which can be even
a binary file if the operation happens in /bin directory, for example.

So, [2] recommends not to use '>' sign as our prompt symbol, because it
can be dangerous and cause serious damages as explained in the blog post
at [1].

Now, there are historical reasons to use '$' as a prompt symbol. One,
is that in UNIX Version 8, there is a variable called stdprompt in
usr/src/cmd/sh/msg.c source file [3] defined as:
char   stdprompt[]  = "$ ";

So, from UNIX Version 8, in the sh shell, the standard prompt string, by
default is "$ ".

As for why the dollar sign '$' was chosen as prompt symbol,
some users at [4] implied it could be a suggestion of the
letters "sh" or even a reminiscent of the word $hell. Who
knows?

Anyway, let's go back to the standard prompt symbol, as shell
in GNU/Linux world is something different from the Command
Prompt on Windows.

Refs:
[0] - https://www.gnu.org/software/bash/manual/html_node/Redirections.html
[1] - https://tanelpoder.com/posts/how-to-stay-safe-in-shell/
[2] - https://ss64.com/bash/syntax-prompt.html
[3] - https://minnie.tuhs.org/cgi-bin/utree.pl?file=V8/usr/src/cmd/sh/msg.c
[4] - https://superuser.com/questions/57575/what-is-the-origin-of-the-unix-dollar-prompt

Signed-off-by: Thiago Farina <tfransosi@gmail.com>