bash: add one whitespace after the prompt symbol
In commit 33993f645c6b ("bash: remove two white spaces from prompt") I removed
this trailing whitespace. But back them, the hostname was bigger and the full path of the working directory was being displayed, leaving very little space
to the command(s) when deep in the directory hierarchy.

But this situation has been changed with the previous changes. The hostname
is just three letters long and we just display the basename of the working
directory.

The main reason to add this trailing whitespace after the prompt symbol back
in my PS1 prompt string is a historical one and because I want to have it
closer to the standard.

The historical one is because the shell sh of UNIX Version 8, had it this way.
At https://minnie.tuhs.org/cgi-bin/utree.pl?file=V8/usr/src/cmd/sh/msg.c,
there is a variable called stdprompt and it is defined as:
char  stdprompt[]   = "$ ";

Another reason is because the PS1 string in the .bashrc installed by default
in Ubuntu also has this trailing whitespace. The string is:
PS1='${debian_chroot:+($debian_chroot)}\[\033[01;32m\]\u@\h\[\033[00m\]:\[\\
033[01;34m\]\w\[\033[00m\]\$ '

The PS1 string in Fedora 11 also has it:
PS1='[\u@\h \W]\\$ '

So, let's have a trailing whitespace in our PS1 string too to keep things
standard as possible.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>