urxvt: revert cursor shape to filled box
In commit 6e227a96995a ("urxvt: change cursor shape to underscore") because
I wanted to make it look like the Command Prompt of Windows XP. The problem
is, I use emacs on the terminal rather than the gui all the time, so
from the screenshot you can see how hard is to find where the cursor is
at any given time as opposed to the time when it is a filled box.

The thing is, in the Command Prompt at Windows XP, the underline or underscore
is way bigger and much more visible, but urxvt does not have anything similar
besides a thin underscore line (which we can barely see on emacs).

Other terminals like gnome-terminal, xfce4-terminal and xterm, all of them
have a filled box as the default cursor shape. Which is also the default
shape in urxvt. So, that is another main reason to revert it to a filled
box.

So, for making easier and fun to edit everything on emacs in the command line
again, let's switch back the cursor to a filled box and no more losing track
of it.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>