bash: adopt Fedora default shell prompt style fully
Fedora 11 has the PS1 variable defined like this:

PS1="[\u@\h \W]\\$ "

Yes, the variable is defined in double quotes and without colors (monochrome).

But the main thing here is the use of the square brackets [] before the prompt
symbol.

This is what our prompt is missing and what we are adding to it right now.

One key difference from it is that our PS1 has a call to the __git_ps1 function
to show the branch name and status when inside a git repository, but that is
fine by me as it does not change the overall look too much.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>