bash: write uncolored version of PS1 string in a single line
Previously, it was written in a multiline fashion because that makes
git blame easier and because the colored version of PS1 was written
this way.

But I plan to switch to a monochrome version (uncolored) and having
it in a single line, although it will make git blame harder, makes
reading it much easier. Also, bash evaluates it all the time, so there
may be subtle performance reasons lying there.

There is a also subtle differences and nuances about single quotes versus
double quotes [0] and [1], but that is topic for another change.

There was no visual changes.

For now, let's go this way.

Refs:
[0] - https://ss64.com/bash/syntax-prompt.html
[1] - https://ss64.com/bash/syntax-quoting.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>