bash: stop exporting PS1 variable
The variable PS1 does not need to be exported into the environment.

Why is that? Because the prompt is only needed in an interactive
shell, as it is there to help the user interact with the shell.

And according to Bash's documentation [0], it is unset in
non-interactive shells.

When the shell is running interactively, Bash will expand and
display PS1 before reading the first line of a command [1].

Refs:
[0] - https://www.gnu.org/software/bash/manual/html_node/Is-this-Shell-Interactive_003f.html
[1] - https://www.gnu.org/software/bash/manual/html_node/Interactive-Shell-Behavior.html
- https://unix.stackexchange.com/questions/247585/to-export-or-not-to-export-bash-ps1-variable
- https://technicqa.com/do-you-need-to-export-ps1-into-the-environment/

Signed-off-by: Thiago Farina <tfransosi@gmail.com>