bash: remove side comments from colored PS1 string
Previously, these comments were there to help explain
the \ special characters. But now, we have these special
characters moved into PS1_XXX variables and having the
documentation near them helps when reading the code.

This opens the door for writing the colored version of
the PS1 string in a single line too.

The info on the special characters was found at
https://www.gnu.org/software/bash/manual/html_node/Controlling-the-Prompt.html.

There was no visual changes this time.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>