bash: use two backslashes to display the right prompt symbol
Bash has these special characters [0] which can appear in the
prompt variable PS1 and it has this one:

\$  If the effective uid is 0, #, otherwise $.

So, for Bash, this \$ has a special meaning when it
appears in PS1, and it should display a # when the user
is root (uid is 0).

The problem is, until now, with our current (or previous) code,
when we run the shell as superuser (root) with `sudo -s`, what
we see is a $ as prompt symbol rather than #.

Why is that? Looks like, it is because when in double quotes,
backslashes are *removed* if they are followed by one of the
following characters: ‘$’, ‘`’, ‘"’, ‘\’, or newline [1].

We can see that seems to be the case:

$ echo "\$"
$

$ echo "\\$"
\$

With one backslash (colored PS1):

$ echo $PS1
[\[\033[0;32m\]\u\[\033[00m\]\[\033[0;32m\]@\h\[\033[00m\] \[\033[0;34m\]\W\[\033[00m\]\[\033[0;33m\]$(__git_ps1)\[\033[00m\]]${PROMPT_SYMBOL}

$ echo ${PROMPT_SYMBOL}
$

Same thing with the uncolored PS1 in tty1 (Ctrl+Alt+F1):

$ echo $PS1
[\u@\h \W$(__git_ps1)]${PROMPT_SYMBOL}

Not good. We define our PS1 in double quotes, and that is
eating the prompt symbol backslash.

The fix? To fix that we need to add one more backslash there.
Because the backslash when followed by another backslash, will
 retain its special meaning.

With two backslashes, what Bash sees and interprets is:

$ echo $PS1
[\[\033[0;32m\]\u\[\033[00m\]\[\033[0;32m\]@\h\[\033[00m\] \[\033[0;34m\]\W\[\033[00m\]\[\033[0;33m\]$(__git_ps1)\[\033[00m\]]\$

In the tty1:

$ echo $PS1
[\u@\h \W$(__git_ps1)]\$

Now, it looks right.

Refs:
[0] - https://www.gnu.org/software/bash/manual/html_node/Controlling-the-Prompt.html
[1] - https://www.gnu.org/software/bash/manual/html_node/Double-Quotes.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>