lightdm: change lightdm-gtk-greeter background to use Fedora 11 Leonidas wallpaper
We changed the desktop wallpaper to Fedora 11 Leonidas wallpaper,
and now we changed lightdm-gtk-greeter background to use the same
wallpaper.

The color of the previous wallpaper from Ubuntu 6.10, matches
the orange tone used by the Human theme there, but not much the
blue color used by the Clearlooks-Phenix theme.

That is why switching to this blueish Fedora wallpaper makes more
sense as it fits better with the overall blue color of the theme
I'm using.

The screenshot was taken going to the tty1 with Ctrl+Alt+F1,
and running:

$ sudo loginscreen

Signed-off-by: Thiago Farina <tfransosi@gmail.com>