xfce: remove frame around the clock
The frame around the clock is something that came
from Windows 95, Windows 98.

With Windows XP the frame around the clock is gone.

More importantly, in Fedora 11 and Ubuntu 9.04, the clock on
gnome-panel does not a have a frame around it, as we can see
in the screenshots of Fedora 11.

So, let's remove the frame from the clock to keep making our
system look more GNU/Linux rather than Windows.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>