xfce: align window title in the center
Previously, the window title was aligned in the left because we were trying
to have the same look and feel of Windows 98, where the title is on the left
right next to the icon.

But that is not how it was in the earlier versions of some GNU/Linux
distributions like Fedora and Ubuntu.

As we can see from the Fedora 11 screenshots, the window title was in the
middle.

So, to keep pushing towards the look and feel of the earlier versions of
Fedora and Ubuntu, let's move the title of the window to the center.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>