xfce: add a copy of keyboard shortcuts configuration file
I'm doing this for two reasons. One is that I want to change
the command run for the Super+L shortcut. I want to call
the exact program used to lock the screen, instead of relying
on what xflock4 wants to do.

Second reason is because having a copy will make it much easier
to restore when reinstalling the system. Otherwise, I would have
to go through the GUI and add all my custom keyboard shortcuts
one by one, and I probably would forget something.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>