urxvt: add default value information for a bunch of options
Added this information to a bunch of resource options that were missing it.

Pretty much all options have this information now. This is good because
if we want to change the value to something else but later on decide to
go back to the original value it is much easier.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>