xfce: move web browser icon one position to the left
This moves the web browser icon launcher one position to
the left, right after the application finder.

Why? Because I want to add an icon launcher for mail
reader, and will fit better if it goes right after the
web browser.

In an order of importance or which came first,
from right to left it is: terminal, file manager, web browser.

Then later it will be: terminal, file manager, mail reader, web browser.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>