emacs: disable blinking to match tty behavior
In GUI frames, Emacs enables `blink-cursor-mode` by default, which
makes the cursor to blink. In contrast, in terminal (tty) frames
the cursor does not blink: urxvt is configured with `URxvt.cursorBlink: false`
and inside Emacs `visible-cursor` is set to nil, resulting in a solid,
non-blinking box.

To make GUI and tty behavior consistent, disable blinking in GUI with:

 (when (display-graphic-p)
    (blink-cursor-mode 0)

This gives a stable, non-blinking box cursor everywhere.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>