add xscreensaver to default MODULES list in setup.sh
Include the xscreensaver directory in the default MODULES array so that
~/.xscreensaver is stowed automatically. This ensures the xscreensaver
configuration is version-controlled and part of the standard dotfiles
setup, without requiring a manual ./setup.sh xscreensaver each time.

Before this change, ~/.xscreensaver existed as a regular file and had to
be moved aside to avoid stow conflicts. After resolving that, the file is
now symlinked to dotfiles/xscreensaver/.xscreensaver as intended.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>